Fix: Lint

+++ spec/rails_helper.rb
@@ -68,8 +68,6 @@ RSpec.configure do |config|
 
   # RSpecテスト後に画像ファイル削除の設定
   config.after(:all) do
-    if Rails.env.test?
-      FileUtils.rm_rf(Dir["#{Rails.root}/public/uploads_#{Rails.env}/"])
-    end
+    FileUtils.rm_rf(Dir["#{Rails.root}/public/uploads_#{Rails.env}/"]) if Rails.env.test?
   end
 end
